requirement.txt
streamlit
Pillow
numpy
opencv-python
ultralytics
torch
